// types/profile.ts
export interface Profile {
    id: string
    username: string | null
    full_name: string | null
    bio: string | null
    avatar_url: string | null
    background_url: string | null
    website: string | null
    location: string | null
    github_url: string | null
    twitter_url: string | null
    created_at: string
}